Move avg_l1_norm
