+++ rl_blox/algorithm/td7.py
@@ -12,44 +12,11 @@ from flax import nnx
 
 from ..blox.double_qnet import ContinuousClippedDoubleQNet
 from ..blox.function_approximator.mlp import MLP
+from ..blox.function_approximator.norm import avg_l1_norm
 from ..blox.function_approximator.policy_head import DeterministicTanhPolicy
 from ..blox.replay_buffer import LAP
 from ..blox.target_net import hard_target_net_update
 from ..logging.logger import LoggerBase
-
-
-def avg_l1_norm(x: jnp.ndarray, eps: float = 1e-8) -> jnp.ndarray:
-    r"""AvgL1Norm.
-
-    A normalization layer that divides the input vector by its average absolute
-    value in each dimension, thus keeping the relative scale of the embedding
-    constant throughout learning.
-
-    Let :math:`x_i` by the i-th dimension of an N-dimensional vector x, then
-
-    .. math::
-
-        AvgL1Norm(x) := \frac{x}{\max(\frac{1}{N} \sum_i |x_i|, \epsilon)},
-
-    with a small constant :math:`\epsilon`.
-
-    AvgL1Norm protects from monotonic growth, but also keeps the scale of the
-    downstream input constant without relying on updating statistics.
-
-    Parameters
-    ----------
-    x : array
-        Input vector(s).
-
-    eps : float
-        Small constant.
-
-    Returns
-    -------
-    avg_l1_norm_x : float
-        AvgL1Norm(x).
-    """
-    return x / jnp.maximum(jnp.mean(jnp.abs(x), axis=-1, keepdims=True), eps)
 
 
 class SALE(nnx.Module):
@@ -1162,11 +1129,9 @@ def train_td7(
             action = env.action_space.sample()
         else:
             key, action_key = jax.random.split(key, 2)
-            action = (
-                np.asarray(
-                    _sample_actions(
-                        fixed_embedding, actor, jnp.asarray(obs), action_key
-                    )
+            action = np.asarray(
+                _sample_actions(
+                    fixed_embedding, actor, jnp.asarray(obs), action_key
                 )
             )
